Update logo branding, prevent file overwrite

+++ src/Commands/Traits/ProcessStubFiles.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace WPGen\Commands\Traits;
 
 /**
@@ -32,14 +33,24 @@ trait ProcessStubFiles
      * @param $stub_path
      * @param $target_path
      * @param $files
+     * @param bool $overwrite
      */
-    protected function processFiles( $stub_path, $target_path, $files ){
+    protected function processFiles( $stub_path, $target_path, $files, $overwrite = false ) {
+
         // Process files.
-        foreach( $files as $file ) {
+        foreach ( $files as $file ) {
+
             // Read source file.
             $data = file_get_contents( $stub_path . $file['source'] );
+
             // Replace values.
             $data = $this->replaceValues( $data );
+
+            // Skip if file already exists.
+            if ( !$overwrite && file_exists( $target_path . $file['target'] ) ) {
+                continue;
+            }
+
             // Store file.
             file_put_contents( $target_path . $file['target'], $data );
         }
